mcfz, model/dr: tracing ignore list and improved handling of COND in driller (#146)
- Add a `spec_level` field to the DR trace, reduce `size` field to `uint8_t`
- Add `--ignorelist` to the DR model and propagate the change to mcfuzz
- Refactor instrumentation state machine to live completely inside of the Dispatcher (see image for reference)
- Improve McFuzz's driller to handle also COND violations

+++ rvzr/model_dynamorio/backend/include/types/file_buffer.hpp
@@ -12,11 +12,9 @@
 #include <array>
 #include <cstdint>
 #include <fstream>
-#include <iostream>
 #include <sstream>
 
-const size_t DEFAULT_FILE_BUFFER_SIZE = 8 * 1024 * 1024; // 8 MB
-
+const uint64_t DEFAULT_FILE_BUFFER_SIZE = 8ULL * 1024 * 1024; // 8 MB
 
 /// @brief A buffer backed by a file: once the buffer reaches a given threshold, it gets
 /// automatically spilled into the backing file. Entries can only be appended to the buffer.
